Move EventType to uefi-raw
