+++ uefi-raw/src/table/boot.rs
@@ -3,6 +3,40 @@
 use crate::{PhysicalAddress, VirtualAddress};
 use bitflags::bitflags;
 
+bitflags! {
+    /// Flags describing the type of an UEFI event and its attributes.
+    #[repr(transparent)]
+    pub struct EventType: u32 {
+        /// The event is a timer event and may be passed to `BootServices::set_timer()`
+        /// Note that timers only function during boot services time.
+        const TIMER = 0x8000_0000;
+
+        /// The event is allocated from runtime memory.
+        /// This must be done if the event is to be signaled after ExitBootServices.
+        const RUNTIME = 0x4000_0000;
+
+        /// Calling wait_for_event or check_event will enqueue the notification
+        /// function if the event is not already in the signaled state.
+        /// Mutually exclusive with `NOTIFY_SIGNAL`.
+        const NOTIFY_WAIT = 0x0000_0100;
+
+        /// The notification function will be enqueued when the event is signaled
+        /// Mutually exclusive with `NOTIFY_WAIT`.
+        const NOTIFY_SIGNAL = 0x0000_0200;
+
+        /// The event will be signaled at ExitBootServices time.
+        /// This event type should not be combined with any other.
+        /// Its notification function must follow some special rules:
+        /// - Cannot use memory allocation services, directly or indirectly
+        /// - Cannot depend on timer events, since those will be deactivated
+        const SIGNAL_EXIT_BOOT_SERVICES = 0x0000_0201;
+
+        /// The event will be notified when SetVirtualAddressMap is performed.
+        /// This event type should not be combined with any other.
+        const SIGNAL_VIRTUAL_ADDRESS_CHANGE = 0x6000_0202;
+    }
+}
+
 newtype_enum! {
 /// Interface type of a protocol interface.
 pub enum InterfaceType: u32 => {
